+++ packages/cargo/jns42-core/src/naming/sentence.rs
@@ -1,3 +1,5 @@
+use crate::ffi::SizedString;
+
 use super::Word;
 use std::{
   iter::{empty, once},
@@ -175,6 +177,69 @@ impl FromIterator<Word> for Sentence {
   fn from_iter<T: IntoIterator<Item = Word>>(iter: T) -> Self {
     Self(iter.into_iter().collect())
   }
+}
+
+/// get the name as camelCase
+#[no_mangle]
+extern "C" fn to_camel_case(value: *const SizedString) -> *const SizedString {
+  assert!(!value.is_null());
+
+  let value = unsafe { &*value };
+  let value = value.as_str();
+
+  let sentence = Sentence::new(value);
+  let result = sentence.to_camel_case();
+  let result = SizedString::new(result);
+  let result = Box::new(result);
+  Box::into_raw(result)
+}
+
+/// get the name as PascalCase
+#[no_mangle]
+extern "C" fn to_pascal_case(value: *const SizedString) -> *const SizedString {
+  assert!(!value.is_null());
+
+  let value = unsafe { &*value };
+  let value = value.as_str();
+
+  let sentence = Sentence::new(value);
+
+  let result = sentence.to_pascal_case();
+  let result = SizedString::new(result);
+  let result = Box::new(result);
+  Box::into_raw(result)
+}
+
+/// get the name as snake_case
+#[no_mangle]
+extern "C" fn to_snake_case(value: *const SizedString) -> *const SizedString {
+  assert!(!value.is_null());
+
+  let value = unsafe { &*value };
+  let value = value.as_str();
+
+  let sentence = Sentence::new(value);
+
+  let result = sentence.to_snake_case();
+  let result = SizedString::new(result);
+  let result = Box::new(result);
+  Box::into_raw(result)
+}
+
+/// get the name as SCREAMING_SNAKE_CASE
+#[no_mangle]
+extern "C" fn to_screaming_snake_case(value: *const SizedString) -> *const SizedString {
+  assert!(!value.is_null());
+
+  let value = unsafe { &*value };
+  let value = value.as_str();
+
+  let sentence = Sentence::new(value);
+
+  let result = sentence.to_screaming_snake_case();
+  let result = SizedString::new(result);
+  let result = Box::new(result);
+  Box::into_raw(result)
 }
 
 #[cfg(test)]
